pyproject.toml
[tool.poetry]
name = "n5challenge"
version = "0.1.0"
description = ""
authors = ["Alejandro Brunacci <Abrunacci87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.3.1"
faker = "^26.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
